add jsdoc, remove makeKeyHoldListener
